flask-backend/requirements.txt
Flask==2.3.2
tensorflow==2.19.0
Pillow
numpy
